Refactor: Move error stack trace code from ZigGlobalObject to JSErrorInfo
This commit extracts all error/stack trace-related code from ZigGlobalObject.cpp
into a new dedicated JSErrorInfo.cpp file for better code organization.

Changes:
- Created JSErrorInfo.h and JSErrorInfo.cpp with all error/stack functions
- Moved Error.captureStackTrace implementation
- Moved Error.prepareStackTrace handling
- Moved computeErrorInfoWithPrepareStackTrace and related functions
- Moved Bun::formatStackTrace and stack formatting utilities
- Moved errorInstanceLazyStackCustomGetter/Setter
- Moved GlobalObject::createCallSitesFromFrames to Zig namespace
- Moved errorConstructorFuncAppendStackTrace
- Moved jsFunctionDefaultErrorPrepareStackTrace
- Moved DEFAULT_ERROR_STACK_TRACE_LIMIT constant
- Updated ZigGlobalObject.h to remove moved function declarations
- Updated ZigGlobalObject.cpp to remove moved implementations
- Added JSErrorInfo.h include to BunProcess.cpp

All error/stack trace functionality remains unchanged, this is purely a
code organization refactor to improve maintainability.
